Fix RT-NEO-DFT copy construction bug

+++ include/realtime.hpp
@@ -104,7 +104,7 @@ namespace ChronusQ {
     IntegrationProgress curState;  ///< Current state of the time propagation
     IntegrationData     data;      ///< Data collection
 
-    size_t printLevel = 1; ///< Amount of printing in RT calc
+    int printLevel = 1; ///< Amount of printing in RT calc
     size_t orbitalPopFreq = 0; ///< Amount of printing in RT calc
     
     bool restart   = false; ///< Restarting calc from bin file
